+++ src/main/java/app/web/pavelk/storage1/services/ClientService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -27,8 +28,9 @@ public class ClientService {
         return clientRepository.findAll(spec);
     }
 
-
+    @Transactional
     public void deleteById(Long id){
+        if (existsById(id))
         clientRepository.deleteById(id);
     }
